Move class Constants to separate package
