+++ src/main/java/org/folio/idmconnect/Constants.java
@@ -1,4 +1,4 @@
-package org.folio.rest.impl;
+package org.folio.idmconnect;
 
 public class Constants {
